pyproject.toml
[project]
name = "capricorn-v"
version = "0.1.0"
description = "原生 Function Calling 驱动的轻量级通用 Agent Runtime — 垂直领域一键扩展"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "Capricorn Contributors"}
]
keywords = ["function-calling", "llm-agent", "agent", "mcp", "ai", "vertical", "capricorn"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3",
]

dependencies = [
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.2.0",
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
    "mcp>=1.0.0",
    "PyYAML>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
]
tiktoken = [
    "tiktoken>=0.5.0",
]

[project.scripts]
capricorn = "run:main"

[tool.setuptools.packages.find]
include = ["agent*", "capabilities*", "config*", "core*", "memory*"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"